MB-17437    CREATE INDEX statement fails in N1QL
Reviewed-on: http://review.couchbase.org/58729
Tested-by: Marco Greco <marco.greco@couchbase.com>

+++ util/hash.go
@@ -17,5 +17,5 @@ func HashString(id string, hashes int) int {
 		h ^= uint(c)
 		h *= 16777619
 	}
-	return int(h) % hashes
+	return int(h % uint(hashes))
 }
